fixes bug with formulas displaying on edit page
